Miskibin uiredesign3 (#17)
* Update breadcrumbs to remove '-results' from segment and improve performance

* feat: Update max-width in clubChart.tsx and page.tsx

This commit updates the max-width property in the `clubChart.tsx` and `page.tsx` files to improve the responsiveness of the club chart component. The max-width value has been changed from 600 to 800, allowing for better display on larger screens. This change ensures that the chart adapts well to different screen sizes and provides a better user experience.

* feat: Add FAQViewSet and related changes

* Update max-width in clubChart.tsx and page.tsx

---------

Co-authored-by: Michal Skibinski <michal.skibinski@nokia.com>