feat: Update server_name in nginx configuration files
This commit updates the `server_name` directive in the `.docker/nginx/default.conf` and `.docker/nginx/default.conf` files. The domain name has been changed from `sejm-stats.com` to `sejm-stats.pl` to match the correct domain for the application. This change ensures that the server correctly handles requests for the updated domain.