Miskibin editor (#19)
* feat: Add ArticleContextViewSet for article context data

This commit adds the ArticleContextViewSet to handle article context data. The viewset includes methods to retrieve lists of envoys, clubs, committees, votings, processes, and interpellations. These lists are used to populate various parts of the application with relevant data. This addition enhances the functionality of the API and improves the user experience by providing comprehensive context information for articles.

* feat: Add ArticleContextViewSet for article context data

* Refactor CreateArticle component to remove unnecessary CSS classes

---------

Co-authored-by: Michal Skibinski <michal.skibinski@nokia.com>